html/semantics/embedded-content-0/media-elements/audio_property_exists.html

<!doctype html>
<html>
 <head>
  <title>video.audio - existence test</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#attr-media-audio">spec reference</a></p>
  <video id="video">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 assert_equals(
   document.getElementById("video").audio,
   undefined,
   "videoElement.audio property exists");
});
  </script>
 </body>
</html>
